refactor(crates): move note parser into charts crate
